Init new Geo provider for geographical and spatial data. Move longitude() and latitude() from Address to Geo
