+++ test/Faker/Provider/AddressTest.php
@@ -15,20 +15,4 @@ class AddressTest extends \PHPUnit_Framework_TestCase
         $faker->addProvider(new Address($faker));
         $this->faker = $faker;
     }
-
-    public function testLatitude()
-    {
-        $latitude = $this->faker->latitude();
-        $this->assertInternalType('float', $latitude);
-        $this->assertGreaterThanOrEqual(-90, $latitude);
-        $this->assertLessThanOrEqual(90, $latitude);
-    }
-
-    public function testLongitude()
-    {
-        $longitude = $this->faker->longitude();
-        $this->assertInternalType('float', $longitude);
-        $this->assertGreaterThanOrEqual(-180, $longitude);
-        $this->assertLessThanOrEqual(180, $longitude);
-    }
 }
